+++ src/app-host/app-host.js
@@ -3,7 +3,6 @@
 var livereload = require('./live-reload-client');
 var Messages = require('messages');
 var telemetry = require('telemetry-helper');
-var xhrProxy = require('./xhr-proxy');
 
 var cordova;
 var oldExec;
@@ -85,7 +84,7 @@ socket.on('init-telemetry', function (data) {
 });
 
 socket.on('init-xhr-proxy', function (data) {
-    xhrProxy.init(); 
+    require('xhr-proxy').init(); 
 });
 
 socket.emit('register-app-host');
