+++ src/modules/xhr-proxy.js
